ARM: dts: imx6ul-kontron-n6x1x-s: Use GPIO1_IO03 as GPIO pin
Currently, GPIO1_IO03 is used by the ADC driver.

On the imx6ull Energy Manager 1.3 board, GPIO1_IO03 is used to read a
digital input value, so move it from the pinctrl_adc1 to pinctrl_gpio,
with the appropriate pull-up.

To read the value of this pin vin userspace:

gpioget 0 3

Signed-off-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>